refactor timestamp component to ui patterns
