+++ apps/studio/components/interfaces/Settings/Logs/LogColumnRenderers/AuthColumnRenderer.tsx
@@ -1,13 +1,9 @@
 import { Column } from 'react-data-grid'
 
 import type { LogData } from '../Logs.types'
-import {
-  RowLayout,
-  SeverityFormatter,
-  TextFormatter,
-  TimestampLocalFormatter,
-} from '../LogsFormatters'
+import { RowLayout, SeverityFormatter, TextFormatter } from '../LogsFormatters'
 import { defaultRenderCell } from './DefaultPreviewColumnRenderer'
+import { TimestampInfo } from 'ui-patterns'
 
 const columns: Column<LogData>[] = [
   {
@@ -20,7 +16,7 @@ const columns: Column<LogData>[] = [
 
       return (
         <RowLayout>
-          <TimestampLocalFormatter value={props.row.timestamp!} />
+          <TimestampInfo value={props.row.timestamp!} />
           {props.row.level && <SeverityFormatter value={props.row.level as string} />}
           <TextFormatter
             className="w-full"
