+++ apps/studio/components/interfaces/Settings/Logs/LogColumnRenderers/DatabaseApiColumnRender.tsx
@@ -1,12 +1,8 @@
 import { Column } from 'react-data-grid'
 import type { LogData } from '../Logs.types'
-import {
-  ResponseCodeFormatter,
-  RowLayout,
-  TextFormatter,
-  TimestampLocalFormatter,
-} from '../LogsFormatters'
+import { ResponseCodeFormatter, RowLayout, TextFormatter } from '../LogsFormatters'
 import { defaultRenderCell } from './DefaultPreviewColumnRenderer'
+import { TimestampInfo } from 'ui-patterns'
 
 const columns: Column<LogData>[] = [
   {
@@ -18,7 +14,7 @@ const columns: Column<LogData>[] = [
       }
       return (
         <RowLayout>
-          <TimestampLocalFormatter value={props.row.timestamp!} />
+          <TimestampInfo value={props.row.timestamp!} />
           <ResponseCodeFormatter row={props} value={props.row.status_code} />
           <TextFormatter className="w-20" value={props.row.method as string} />
           <TextFormatter className="w-full" value={props.row.path as string} />
