+++ apps/studio/components/interfaces/Settings/Logs/LogColumnRenderers/DatabasePostgresColumnRender.tsx
@@ -1,12 +1,8 @@
 import { Column } from 'react-data-grid'
 import type { LogData } from '../Logs.types'
-import {
-  RowLayout,
-  SeverityFormatter,
-  TextFormatter,
-  TimestampLocalFormatter,
-} from '../LogsFormatters'
+import { RowLayout, SeverityFormatter, TextFormatter } from '../LogsFormatters'
 import { defaultRenderCell } from './DefaultPreviewColumnRenderer'
+import { TimestampInfo } from 'ui-patterns'
 
 const columns: Column<LogData>[] = [
   {
@@ -18,7 +14,7 @@ const columns: Column<LogData>[] = [
       }
       return (
         <RowLayout>
-          <TimestampLocalFormatter value={props.row.timestamp!} />
+          <TimestampInfo value={props.row.timestamp!} />
           <SeverityFormatter value={props.row.error_severity as string} />
           <TextFormatter className="w-full" value={props.row.event_message} />
         </RowLayout>
