+++ apps/studio/components/interfaces/Settings/Logs/LogColumnRenderers/DefaultPreviewColumnRenderer.tsx
@@ -1,10 +1,11 @@
 import { Column, RenderCellProps } from 'react-data-grid'
 import type { LogData } from '../Logs.types'
-import { RowLayout, TextFormatter, TimestampLocalFormatter } from '../LogsFormatters'
+import { RowLayout, TextFormatter } from '../LogsFormatters'
+import { TimestampInfo } from 'ui-patterns'
 
 export const defaultRenderCell = (props: RenderCellProps<LogData, unknown>) => (
   <RowLayout>
-    <TimestampLocalFormatter value={props.row.timestamp!} />
+    <TimestampInfo value={props.row.timestamp!} />
     <TextFormatter className="w-full" value={props.row.event_message} />
   </RowLayout>
 )
