+++ packages/ui-patterns/index.tsx
@@ -22,3 +22,4 @@ export * from './SqlToRest'
 export * from './PromoToast'
 export * from './admonition'
 export * from './ComputeBadge/ComputeBadge'
+export * from './TimestampInfo/TimestampInfo'
