+++ packages/ui-patterns/package.json
@@ -14,6 +14,7 @@
     "cmdk": "^1.0.0",
     "common": "*",
     "common-tags": "^1.8.2",
+    "dayjs": "^1.11.13",
     "framer-motion": "^11.1.9",
     "lodash": "*",
     "lucide-react": "*",
